tilda.c: wayland dbus visibility toggler
in order to toggle visibility in Wayland, commands
have to be sent via dbus. This commit introduces
command line switch & function to call dbus method.
It changes the order of tilda startup, as a window
is not needed, if we intend to exit right after sending cmd.
Lastly, there are a couple of formatting fixes, and readme entry.

Signed-off-by: Krzysztof Królczyk <Krzysztof.Krolczyk@o2.pl>